Removing the `_id` column and making `entryid` primary key. Moved the `TasksLocalDataSourceTest` to correct package.
