+++ todoapp/app/src/main/java/com/example/android/architecture/blueprints/todoapp/data/source/local/TasksPersistenceContract.java
@@ -29,7 +29,7 @@ public final class TasksPersistenceContract {
 
     /* Inner class that defines the table contents */
     public static abstract class TaskEntry implements BaseColumns {
-        public static final String TABLE_NAME = "task";
+        public static final String TABLE_NAME = "tasks";
         public static final String COLUMN_NAME_ENTRY_ID = "entryid";
         public static final String COLUMN_NAME_TITLE = "title";
         public static final String COLUMN_NAME_DESCRIPTION = "description";
